[WEB-3216] - cleanup
